Database Names
Creation of a "database" to store the names of all the elements that will appear in the scroll view. The scriptable objects tool was used to have a quick and practical way of accessing and generating random names for the scroll view.

Added DatabaseSO.cs
Added Database Element Name.asset